simplifies world get methods and system manager add/remove/get methods
to keep the API as simple and error free as possible, some functionality has been
removed in favor of a clean and intuitive API.

should the need for a more parameterized API arise, we will consider reverting back
at that point
